SpreadsheetMetadataColorPickerComponent moved to metadatacolorpicker package
